[#25] Refactor: Update package names in design system
This commit updates the package names within the design system module from `ir.vahid.core.design_system` to `ir.vahid.core.designsystem`, including the `Color`, `Type`, and `Theme` files. It also updates the import in `MainActivity` accordingly.
