+++ core/design-system/src/main/java/ir/vahid/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package ir.vahid.core.design_system.theme
+package ir.vahid.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
