+++ core/screenshot-testing/src/main/java/ir/vahid/core/testing/util/ScreenshotHelper.kt
@@ -27,7 +27,7 @@ import com.google.android.apps.common.testing.accessibility.framework.Accessibil
 import com.google.android.apps.common.testing.accessibility.framework.AccessibilityViewCheckResult
 import com.google.android.apps.common.testing.accessibility.framework.integrations.espresso.AccessibilityViewCheckException
 import com.google.android.apps.common.testing.accessibility.framework.utils.contrast.BitmapImage
-import ir.vahid.core.design_system.theme.FrameworkTheme
+import ir.vahid.core.designsystem.theme.FrameworkTheme
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers
 import org.robolectric.RuntimeEnvironment
